+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.2.0] - 2020-??-??
+### Added
+- Added feedback when you vote for an element and it gets voted through.
+
 ### Fixed
+- Made sure adding elements with the same name as others properly adds the element. (#1)
 - Fixed not being able to edit an element's name after setting it to a blank string. (#2)
 
 ## [1.1.0] - 2020-05-31